roachtest: move or recovery test into or file
Epic: None

Release note: None
