+++ pkg/cmd/roachtest/tests/backup_restore_roundtrip.go
@@ -20,7 +20,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/roachtestutil/clusterupgrade"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/spec"
 	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/test"
-	"github.com/cockroachdb/cockroach/pkg/jobs/jobspb"
 	"github.com/cockroachdb/cockroach/pkg/roachprod/install"
 	"github.com/cockroachdb/cockroach/pkg/roachprod/logger"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
@@ -109,25 +108,6 @@ func registerBackupRestoreRoundTrip(r registry.Registry) {
 			},
 		})
 	}
-}
-
-func registerOnlineRestoreRecovery(r registry.Registry) {
-	r.Add(registry.TestSpec{
-		Name:                       "backup-restore/online-restore-recovery",
-		Timeout:                    4 * time.Hour,
-		Owner:                      registry.OwnerDisasterRecovery,
-		Cluster:                    r.MakeClusterSpec(4, spec.WorkloadNode()),
-		EncryptionSupport:          registry.EncryptionMetamorphic,
-		NativeLibs:                 registry.LibGEOS,
-		CompatibleClouds:           registry.Clouds(spec.GCE, spec.Local),
-		Suites:                     registry.Suites(registry.Nightly),
-		TestSelectionOptOutSuites:  registry.Suites(registry.Nightly),
-		Randomized:                 true,
-		RequiresDeprecatedWorkload: true, // uses schemachange
-		Run: func(ctx context.Context, t test.Test, c cluster.Cluster) {
-			testOnlineRestoreRecovery(ctx, t, c)
-		},
-	})
 }
 
 // backup-restore/round-trip tests that a round trip of creating a backup and
@@ -456,146 +436,4 @@ func createDriversForBackupRestore(
 		return nil, nil, nil, err
 	}
 	return d, runBackgroundWorkload, tables, nil
-}
-
-func testOnlineRestoreRecovery(ctx context.Context, t test.Test, c cluster.Cluster) {
-	testRNG, seed := randutil.NewLockedPseudoRand()
-	workloadSeed := testRNG.Int63()
-	t.L().Printf("random seed: %d", seed)
-
-	c.Start(ctx, t.L(), roachtestutil.MaybeUseMemoryBudget(t, 50), install.MakeClusterSettings(), c.CRDBNodes())
-	m := c.NewDeprecatedMonitor(ctx, c.CRDBNodes())
-	const jobStatusWait = time.Minute * 5
-
-	m.Go(func(ctx context.Context) error {
-		testUtils, err := setupBackupRestoreTestUtils(
-			ctx, t, c, m, testRNG,
-			withOnlineRestore(true), withCompaction(true),
-		)
-		if err != nil {
-			return err
-		}
-		defer testUtils.CloseConnections()
-
-		dbs := []string{"bank", "tpcc", schemaChangeDB}
-		d, runBackgroundWorkload, _, err := createDriversForBackupRestore(
-			ctx, t, c, m, testRNG, workloadSeed, testUtils, dbs,
-		)
-		if err != nil {
-			return err
-		}
-
-		stopBackgroundCommands, err := runBackgroundWorkload()
-		if err != nil {
-			return err
-		}
-		defer stopBackgroundCommands()
-
-		dbConn := c.Conn(ctx, t.L(), c.CRDBNodes().RandNode()[0])
-		defer dbConn.Close()
-
-		allNodes := labeledNodes{Nodes: c.CRDBNodes(), Version: clusterupgrade.CurrentVersion().String()}
-		bspec := backupSpec{
-			Plan:    allNodes,
-			Execute: allNodes,
-		}
-
-		// We set this pausepoint so that the download job is paused before it
-		// begins downloading external files. This allows us to guarantee that when
-		// we delete an SST file, it won't have already been downloaded by the
-		// download phase and therefore not trigger a failure.
-		// We also must set this BEFORE taking backups. In the event of a full
-		// cluster backup, the link phase of OR will reset the pausepoints back to
-		// the point of the backup, which means this pausepoint will be wiped if it
-		// is set after the backups.
-		if _, err := dbConn.ExecContext(
-			ctx, "SET CLUSTER SETTING jobs.debug.pausepoints = 'restore.before_download'",
-		); err != nil {
-			return errors.Wrap(err, "failed to set pausepoint for online restore")
-		}
-
-		t.L().Printf("starting backup")
-		collection, err := d.createBackupCollection(
-			ctx, t.L(), t, testRNG, bspec, bspec, "online-restore-recovery-backup",
-			true /* internalSystemsJobs */, false, /* isMultitenant */
-		)
-		if err != nil {
-			return err
-		}
-
-		// If we're running a cluster backup, we need to reset the cluster
-		// to restore it. We also intentionally stop background commands so
-		// the workloads don't report errors.
-		if _, ok := collection.btype.(*clusterBackup); ok {
-			t.L().Printf("resetting cluster before restoring full cluster backup")
-			stopBackgroundCommands()
-			expectDeathsFn := func(n int) {
-				m.ExpectDeaths(int32(n))
-			}
-
-			// Between each reset grab a debug zip from the cluster.
-			zipPath := fmt.Sprintf("debug-%d.zip", timeutil.Now().Unix())
-			if err := testUtils.cluster.FetchDebugZip(ctx, t.L(), zipPath); err != nil {
-				t.L().Printf("failed to fetch a debug zip: %v", err)
-			}
-			if err := testUtils.resetCluster(
-				ctx, t.L(), clusterupgrade.CurrentVersion(), expectDeathsFn, []install.ClusterSettingOption{},
-			); err != nil {
-				return err
-			}
-		}
-
-		t.L().Printf("performing online restore of backup")
-		if _, _, err := d.runRestore(
-			ctx, t.L(), testRNG, collection,
-			false /* checkFiles */, true /* internalSystemJobs */, nil, /* mvHelper */
-		); err != nil {
-			return err
-		}
-		downloadJobID, err := d.getORDownloadJobID(ctx, t.L(), testRNG)
-		if err != nil {
-			return err
-		}
-		if err := WaitForPaused(ctx, dbConn, jobspb.JobID(downloadJobID), jobStatusWait); err != nil {
-			return errors.Wrapf(
-				err, "waiting for download job %v to reach paused state", downloadJobID,
-			)
-		}
-
-		// defaultdb is going to be set offline by the failed download job, so we
-		// need to switch to the system database first to avoid any errors.
-		if _, err := dbConn.ExecContext(ctx, "USE system"); err != nil {
-			return err
-		}
-
-		if _, err := dbConn.ExecContext(
-			ctx, "SET CLUSTER SETTING jobs.debug.pausepoints = ''",
-		); err != nil {
-			return errors.Wrap(err, "failed to remove pausepoint for online restore")
-		}
-
-		if err := d.deleteSSTFromBackupLayers(ctx, t.L(), dbConn, collection); err != nil {
-			return err
-		}
-		if _, err := dbConn.ExecContext(ctx, "RESUME JOB $1", downloadJobID); err != nil {
-			return errors.Wrapf(
-				err, "failed to resume download job %v after deleting sst", downloadJobID,
-			)
-		}
-		if err := WaitForResume(ctx, dbConn, jobspb.JobID(downloadJobID), jobStatusWait); err != nil {
-			return errors.Wrapf(
-				err, "waiting for download job %v to reach resumed state", downloadJobID,
-			)
-		}
-		if err := WaitForFailed(ctx, dbConn, jobspb.JobID(downloadJobID), 10*time.Minute); err != nil {
-			return errors.Wrapf(
-				err, "waiting for download job %v to reach failed state", downloadJobID,
-			)
-		}
-		return errors.Wrapf(
-			checkNoExternalBytesRemaining(ctx, dbConn),
-			"cluster did not cleanup all external files after download phase failure",
-		)
-	})
-	m.Wait()
 }
